requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
requests
supabase
python-dotenv
passlib
PyJWT


